+++ spec/pio/open_flow13/match_spec.rb
@@ -403,6 +403,248 @@ describe Pio::OpenFlow13::Match do
       And { match.match_fields[2].masked? == false }
       And { match.match_fields[2].oxm_length == 1 }
     end
+
+    context 'with ether_type: 0x0806, arp_op: 1' do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_op: 1
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_op == 1 }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 16 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 16 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpOp::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == false }
+      And { match.match_fields[1].oxm_length == 2 }
+    end
+
+    context "with ether_type: 0x0806, arp_sender_protocol_address: '1.2.3.4'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_sender_protocol_address: '1.2.3.4'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_sender_protocol_address == '1.2.3.4' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 18 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpSenderProtocolAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == false }
+      And { match.match_fields[1].oxm_length == 4 }
+    end
+
+    context "with ether_type: 0x0806, arp_sender_protocol_address: '1.2.3.4', arp_sender_protocol_address_mask: '255.255.0.0'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_sender_protocol_address: '1.2.3.4',
+          arp_sender_protocol_address_mask: '255.255.0.0'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_sender_protocol_address == '1.2.3.4' }
+      Then { match.arp_sender_protocol_address_mask == '255.255.0.0' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 22 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpSenderProtocolAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == true }
+      And { match.match_fields[1].oxm_length == 8 }
+    end
+
+    context "with ether_type: 0x0806, arp_target_protocol_address: '1.2.3.4'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_target_protocol_address: '1.2.3.4'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_target_protocol_address == '1.2.3.4' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 18 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpTargetProtocolAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == false }
+      And { match.match_fields[1].oxm_length == 4 }
+    end
+
+    context "with ether_type: 0x0806, arp_target_protocol_address: '1.2.3.4', arp_target_protocol_address_mask: '255.255.0.0'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_target_protocol_address: '1.2.3.4',
+          arp_target_protocol_address_mask: '255.255.0.0'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_target_protocol_address == '1.2.3.4' }
+      Then { match.arp_target_protocol_address_mask == '255.255.0.0' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 22 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpTargetProtocolAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == true }
+      And { match.match_fields[1].oxm_length == 8 }
+    end
+
+    context "with ether_type: 0x0806, arp_sender_hardware_address: '11:22:33:44:55:66'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_sender_hardware_address: '11:22:33:44:55:66'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_sender_hardware_address == '11:22:33:44:55:66' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 20 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpSenderHardwareAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == false }
+      And { match.match_fields[1].oxm_length == 6 }
+    end
+
+    context "with ether_type: 0x0806, arp_sender_hardware_address: '11:22:33:44:55:66', arp_sender_hardware_address_mask: 'ff:ff:ff:00:00:00'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_sender_hardware_address: '11:22:33:44:55:66',
+          arp_sender_hardware_address_mask: 'ff:ff:ff:00:00:00'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_sender_hardware_address == '11:22:33:44:55:66' }
+      Then { match.arp_sender_hardware_address_mask == 'ff:ff:ff:00:00:00' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 32 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 26 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpSenderHardwareAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == true }
+      And { match.match_fields[1].oxm_length == 12 }
+    end
+
+    context "with ether_type: 0x0806, arp_target_hardware_address: '11:22:33:44:55:66'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_target_hardware_address: '11:22:33:44:55:66'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_target_hardware_address == '11:22:33:44:55:66' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 24 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 20 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpTargetHardwareAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == false }
+      And { match.match_fields[1].oxm_length == 6 }
+    end
+
+    context "with ether_type: 0x0806, arp_target_hardware_address: '11:22:33:44:55:66', arp_target_hardware_address_mask: 'ff:ff:ff:00:00:00'" do
+      When(:match) do
+        Pio::OpenFlow13::Match.new(
+          ether_type: 0x0806,
+          arp_target_hardware_address: '11:22:33:44:55:66',
+          arp_target_hardware_address_mask: 'ff:ff:ff:00:00:00'
+        )
+      end
+      Then { match.ether_type == 0x0806 }
+      Then { match.arp_target_hardware_address == '11:22:33:44:55:66' }
+      Then { match.arp_target_hardware_address_mask == 'ff:ff:ff:00:00:00' }
+      And { match.class == Pio::OpenFlow13::Match }
+      And { match.length == 32 }
+      And { match.match_type == Pio::OpenFlow13::MATCH_TYPE_OXM }
+      And { match.match_length == 26 }
+      And { match.match_fields.size == 2 }
+      And do
+        match.match_fields[1].oxm_class ==
+          Pio::OpenFlow13::Match::OXM_CLASS_OPENFLOW_BASIC
+      end
+      And do
+        match.match_fields[1].oxm_field ==
+          Pio::OpenFlow13::Match::ArpTargetHardwareAddress::OXM_FIELD
+      end
+      And { match.match_fields[1].masked? == true }
+      And { match.match_fields[1].oxm_length == 12 }
+    end
   end
 
   def read_raw_data_file(name)